add percentage refactoring term()

+++ Gemwest/CCalculator.h
@@ -6,8 +6,8 @@ using namespace System::Diagnostics;
 ref class CCalculator abstract
 {
 public:
-	//virtual Decimal term() const = 0;
-	virtual Decimal term()  = 0;
+	//virtual Double term() const = 0;
+	virtual Double term()  = 0;
 	virtual Decimal recut_weight() = 0;
 	//virtual String^ print() {
 	//	return System::Convert::ToString(5+"many ct");
