+++ Gemwest/CDround.h
@@ -21,10 +21,10 @@ public:
 		Boolean recut, //
 
 		Boolean depthisperc) : CDcalc(cutName, fac, d1, d2, dep, sg, adj, gt, pv, so, isdiamond, interp, recut, depthisperc) {}
-    //virtual Decimal term() const = 0;
-    Decimal term() override;
+    //virtual Double term() const = 0;
+    Double term() override;
     Decimal recut_weight() override;
-    //virtual Decimal term() override;
+    //virtual Double term() override;
     //String^ recut_weight() override;
 
 };
